requirements.txt
# Core Streamlit Framework
streamlit>=1.28.0,<2.0.0
streamlit-aggrid>=0.3.4,<1.0.0

# Data Models and Validation
pydantic>=2.0.0,<3.0.0

# Machine Learning Libraries
scikit-learn>=1.3.0,<2.0.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0

# Visualization Libraries
plotly>=5.15.0,<6.0.0
matplotlib>=3.7.0,<4.0.0

# LLM and AI Libraries
langchain>=0.0.300,<1.0.0
openai>=1.0.0,<2.0.0

# Database (SQLite is built-in with Python)
# Additional database drivers if needed
# psycopg2-binary>=2.9.0  # PostgreSQL
# pymysql>=1.1.0          # MySQL

# Utility Libraries
python-dotenv>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
urllib3>=1.26.0,<3.0.0

# Development Dependencies (Optional)
# pytest>=7.4.0,<8.0.0
# black>=23.0.0,<24.0.0
# flake8>=6.0.0,<7.0.0
# mypy>=1.5.0,<2.0.0

# Optional Performance Enhancements
# redis>=4.6.0,<5.0.0     # For caching
# celery>=5.3.0,<6.0.0    # For background tasks

# Optional Advanced Features
# spacy>=3.6.0,<4.0.0     # For NLP
# transformers>=4.30.0,<5.0.0  # For Hugging Face models
# torch>=2.0.0,<3.0.0     # For PyTorch models

# Security and Authentication (Optional)
# cryptography>=41.0.0,<42.0.0
# pyjwt>=2.8.0,<3.0.0

# Monitoring and Logging (Optional)
# sentry-sdk>=1.30.0,<2.0.0
# loguru>=0.7.0,<1.0.0